refactor embeds
