+++ src/startup/mini-embed.tsx
@@ -11,7 +11,7 @@ import "../adapter";
 import dayjs from "dayjs";
 import relativeTime from "dayjs/plugin/relativeTime";
 import { createRoot } from "react-dom/client";
-import { MiniRunScreen } from "~/screens/mini-run";
+import { MiniRunScreen } from "~/screens/mini-embed";
 import { openConnection } from "~/screens/surrealist/connection/connection";
 import { startConfigSync } from "~/util/config";
 import { handleWindowMessage } from "~/util/messaging";
