+++ tools/mini-embed.html
@@ -9,7 +9,7 @@
 
 	<link rel="shortcut icon" href="/favicon.ico" type="image/x-icon" />
 
-	<title>Surrealist Mini Generator</title>
+	<title>Surrealist Mini</title>
 
 	<!-- Twitter Cards & Open Graph -->
 	<meta name="description" content="Surrealist is the ultimate way to visually manage your SurrealDB database">
@@ -23,11 +23,12 @@
 	<meta name="og:description" content="Surrealist is the ultimate way to visually manage your SurrealDB database">
 	<meta name="og:image"
 		content="https://raw.githubusercontent.com/surrealdb/surrealist/main/.github/images/thumbnail.jpg">
+	<meta name="surrealist-mini">
 </head>
 
 <body>
 	<div id="root"></div>
-	<script src="/src/startup/mini-new.tsx" type="module"></script>
+	<script src="/src/startup/mini-embed.tsx" type="module"></script>
 </body>
 
 </html>
