--- tools/mini-run.html
@@ -1,34 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-	<meta charset="utf-8" />
-	<meta name="theme-color" content="#d500c6" />
-	<meta content="width=device-width, initial-scale=1.0" name="viewport" />
-	<meta name="author" content="SurrealDB Ltd">
-
-	<link rel="shortcut icon" href="/favicon.ico" type="image/x-icon" />
-
-	<title>Surrealist Mini</title>
-
-	<!-- Twitter Cards & Open Graph -->
-	<meta name="description" content="Surrealist is the ultimate way to visually manage your SurrealDB database">
-	<meta name="twitter:card" content="summary_large_image">
-	<meta name="twitter:site" content="@SurrealDB">
-	<meta name="twitter:title" content="Surrealist Mini">
-	<meta name="twitter:image"
-		content="https://raw.githubusercontent.com/surrealdb/surrealist/main/.github/images/thumbnail.jpg">
-	<meta name="og:title" content="Surrealist Mini">
-	<meta name="og:type" content="website">
-	<meta name="og:description" content="Surrealist is the ultimate way to visually manage your SurrealDB database">
-	<meta name="og:image"
-		content="https://raw.githubusercontent.com/surrealdb/surrealist/main/.github/images/thumbnail.jpg">
-	<meta name="surrealist-mini">
-</head>
-
-<body>
-	<div id="root"></div>
-	<script src="/src/startup/mini-run.tsx" type="module"></script>
-</body>
-
-</html>
